requirements.txt
alembic==1.7.1
asgiref==3.4.1
blinker==1.4
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==8.0.1
colorama==0.4.3
distro==1.4.0
entrypoints==0.3
fasteners==0.14.1
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.0
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
libpq-dev==13.4
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
monotonic==1.5
MouseInfo==0.1.3
netifaces==0.10.4
numpy==1.21.1
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
protobuf==3.6.1
psycopg2==2.9.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyDrive==1.3.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyperclip==1.8.2
PyRect==0.1.4
pyRFC3339==1.1
PyScreeze==0.1.27
python-dateutil==2.7.3
python-debian===0.1.36
python-decouple==3.4
python3-xlib==0.15
PyTweening==1.0.3
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
soupsieve==2.2.1
SQLAlchemy==1.4.22
sqlparse==0.4.1
systemd-python==234
uritemplate==0.6
urllib3==1.25.8
wadllib==1.3.3
Werkzeug==2.0.1
zipp==3.5.0
